Restyled and Prettier fix
The .prettierrc.json file was getting flagged because it was all fucked up. That file is supposed to be formatted as shown in this documentation https://prettier.io/docs/en/configuration.html so I think I fixed that. I also added the config file for Restyled that was supposed to be in the root of the repo too. The config file for that is configured with the default configs for Restyler as seen here https://github.com/restyled-io/restyler/blob/master/config/default.yaml more info about the configs can also be seen at https://github.com/restyled-io/restyled.io/wiki/Configuring-Restyled