Move master > release (#55)
* New 8-ball functionality.

Added gifs to go along with responses. Also moved the messages into 8-ball.json.

* Formatting

Forgot to format it, oopsie

* debug shit

This should return what link was used in kiss action. Some are broken and it would be epic to see which ones are.

* Json list update

Went through links and figured out which were dead for the kiss action.

* Removed debug code

* BE GONE DEAD LINKS

Updated actions.json and removed bad links

* typos in it yikes

* Added new Firebase commands

Signed-off-by: Sean Doran <sdoran35@gmail.com>

* Restyled and Prettier fix

The .prettierrc.json file was getting flagged because it was all fucked up. That file is supposed to be formatted as shown in this documentation https://prettier.io/docs/en/configuration.html so I think I fixed that. I also added the config file for Restyled that was supposed to be in the root of the repo too. The config file for that is configured with the default configs for Restyler as seen here https://github.com/restyled-io/restyler/blob/master/config/default.yaml more info about the configs can also be seen at https://github.com/restyled-io/restyled.io/wiki/Configuring-Restyled

* link update

a.safe.moe is dead rip in peace. replaced with catbox.moe

* formatting update of everything

WEW LAD THAT IS A LOT OF CHANGED FILES. Everything got changed to use the style listed in the .prettierrc.json file. So now tab width is a standard 4 spaces, trailing commas are a thing when needed, semicolons are standard now(you are welcome Sean <3), single quotes are preferred, and bracket spacing is standard now(default setting with Prettier). More info about the settings can be found here https://prettier.io/docs/en/options.html

* Restyler change

Fuck you Restyler, just use prettier and stop being dumb. This should make it so it just uses Prettier and our configs.

* Created .prettierignore

Currently set up to ignore .yaml files since it doesn't like .restyled.yaml and the tabs for the bottom part of it.  Also set to ignore certain folders hopefully

* shutdown update

This should check if the bot is in any voice channels? I cant test it so who knowsssssssss

* formatting shutdown

HEY WILLIAM DONT FORGET TO FORMAT SHIT YOU DUMBASS

* 8ball update poggers

made 8ball a bit more simplistic and hopefully fixed the gif problem.

* formatting

* AHAHSDJKHASKDJHASD

* Updated commands with new throttling.

* Updated commands with new throttling.

* Updated commands with new throttling.

* Updated commands with new throttling.

* grope got simplified pog

* hand got simplified and prettier ran again pog

* Wow hug, kiss, lick, and nobully are better. That is pretty pog. It's also 114 at night and I am tired. Please send help.

* Update package.json

* fixed pkg.json

* Moved JSON files into correct direcory.

Co-authored-by: Snoo <wm28744@gmail.com>